--- product/static/product_list/product_list.css
@@ -1,19 +1,0 @@
-
-.name_black{
-    color: black; 
-    font-size: 24px; 
-    font-family: Inknut Antiqua; 
-    font-weight: 400; 
-    letter-spacing: 0.20px; 
-    word-wrap: break-word
-}
-
-.name_color{
-    color: #0ECEDA; 
-    font-size: 24px; 
-    font-family: Inknut Antiqua; 
-    font-weight: 400; 
-    letter-spacing: 0.20px; 
-    word-wrap: break-word"
-}
-
