--- product/templates/product/navbar.html
@@ -1,68 +1,0 @@
-<!DOCTYPE html>
-{% load static %}
-<html lang="ko">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-    
-    <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.9.2/dist/umd/popper.min.js" ></script>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.min.js" integrity="sha384-cVKIPhGWiC2Al4u+LWgxfKTRIcfu0JTxR+EQDz/bgldoEyl4H0zUF0QKbrJ0EcQF" crossorigin="anonymous"></script>
-
-    <link rel="stylesheet" href="{% static 'product_list/product_list.css' %}" type="text/css" media="screen">
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.2/css/all.min.css" integrity="sha512-SnH5WK+bZxgPHs44uWIX+LLJAJ9/2PkPKZ5QiAj6Ta86w+fsb2TkcmfRyVX3pBnMFcV7oQPJkl9QevSCWr3W6A==" crossorigin="anonymous" referrerpolicy="no-referrer" />
-</head>
-<body>
-  <nav class="navbar navbar-expand-lg navbar-white bg-white mb-3">
-    <div class="container">
-        <a class="navbar-brand" href="#!"><span class="name_black">S</span><span class="name_color">E</span><span class="name_black">LLER</span></a>
-        <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation"><span class="navbar-toggler-icon"></span></button>
-        <div class="collapse navbar-collapse" id="navbarSupportedContent">
-            <div class="collapse navbar-collapse col-la " style="margin-left: 200px;" id="navbarSupportedContent">
-                <form class="d-flex" role="search" style="width: 489px;">
-                  <input class="form-control me-2" type="search" placeholder="Search" aria-label="Search">
-                  <button class="btn btn-outline-success" type="submit"><i class="fa-solid fa-magnifying-glass"></i></button>
-                </form>
-            </div>
-            <ul class="navbar-nav ms-auto mb-2 mb-lg-0 mt-2">
-                <li class="nav-item"><a class="nav-link d-flex flex-column align-items-center" href="#"><i class="fa-solid fa-book-open "></i><div style="font-size: 11px;">주문내역</div></a></li>
-                <li class="nav-item"><a class="nav-link d-flex flex-column align-items-center" href="#!"><i class="fa-solid fa-cart-shopping"></i><div style="font-size: 11px;">장바구니</div></a></li>
-                <li class="nav-item dropdown">
-                  <div class="nav-link dropdown-toggle d-flex flex-column align-items-center" id="navbarDropdown" role="button" data-bs-toggle="dropdown" aria-expanded="false"><i class="fa-regular fa-user "></i><div style="font-size: 12px;">내정보</div></div>
-                  <ul class="dropdown-menu" aria-labelledby="navbarDropdown">
-                      <li><a class="dropdown-item" href="#!">개발예정</a></li>
-                      <li><hr class="dropdown-divider" /></li>
-                      <li><a class="dropdown-item" href="#!">개발예정</a></li>
-                      <li><a class="dropdown-item" href="#!">개발예정</a></li>
-                  </ul>
-                </li>
-                
-            </ul>
-        </div>
-    </div>
-  </nav> 
-
-
-  <nav class="navbar navbar-expand-lg navbar-white bg-white">
-    <div class="container">
-        
-        <div class="collapse navbar-collapse row container" id="navbarSupportedContent">
-            
-                <ul class="navbar-nav ms-auto mb-2 mb-lg-0" style="display: flex; justify-content: space-between;">
-                    <li class="nav-item " ><a class="nav-link" href="/product/category/식품/">식품</a></li>
-                    <li class="nav-item "><a class="nav-link" href="/product/category/패션주얼리/">패션/주얼리</a></li>
-                    <li class="nav-item "><a class="nav-link" href="/product/category/홈리빙/">홈리빙</a></li>
-                    <li class="nav-item "><a class="nav-link" href="/product/category/케이스문구/">케이스/문구</a></li>
-                    <li class="nav-item "><a class="nav-link" href="/product/category/뷰티/">뷰티</a></li>
-                    <li class="nav-item "><a class="nav-link" href="/product/category/반려동물/">반려동물</a></li>
-                    <li class="nav-item "><a class="nav-link" href="/product/category/공예/">공예</a></li>
-                </ul>
-            
-        </div>
-    </div>
-  </nav> 
-<hr>  
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
-</body>
-</html>
