--- product/templates/product/product_detail.html
@@ -1,79 +1,0 @@
-{% extends 'product/base.html '%}
-{% block main_area %}
-    <div class="container px-4 px-lg-5 my-5">
-        <div class="row gx-4 gx-lg-5">
-            <div class="col-md-6" >
-                <img class="card-img-top mb-5 " src="https://picsum.photos/seed/{{ product.id }}/800/200" alt="">
-            </div>
-            <div class="col-md-6 d-flex flex-column" >
-                <div class="small mb-1">브랜드명</div>
-                <h1 class="display-5 fw-bolder">{{product.product_name}}</h1>
-                <div class="fs-5 mb-5">
-                    <span>{{product.price}}원</span>
-                </div>
-                <p class="lead">
-                    {{product.description}} 
-                    
-                </p>
-                <div class="mt-auto" name="this">
-                    <div class="bg-light d-flex justify-content-between">
-                        <span class="mx-2">
-                            <input type="hidden" id="csrfToken" value="{{csrf_token}}">
-                            <button type="button" onclick="aa(this)" class="btn btn-success">-</button>
-                            <span id="currentQuantity">1</span>
-                            <button type="button" onclick="aa(this)" class="btn btn-danger">+</button>
-                        </span>
-                        <span id="currentPrice">{{product.price}}원</span>
-                    </div>
-                    <div>
-                        <div class="d-grid gap-2">
-                            <button class="btn btn-info text-light" type="button">장바구니</button>
-                            <button class="btn btn-info text-light" type="button" onclick="modifyCartAmount({{product.pk}},document.getElementById('currentQuantity').textContent )">구매하기</button>
-                        </div>
-                    </div>
-                    
-                </div>
-            </div> 
-        </div>
-    </div>
-    <div class="container">
-        <div class="fs-3 fw-bold">작품정보</div>
-    </div>
-    <script>
-        function getCsrfToken(){
-            return document.getElementById('csrfToken').value;
-        }
-        function modifyCartAmount(product_id, change){
-            var xhr = new XMLHttpRequest();
-                xhr.open('POST', '{% url "product:modify_cart" %}', true);
-                xhr.setRequestHeader('Content-Type', 'application/x-www-form-urlencoded');
-                xhr.setRequestHeader('X-CSRFToken', getCsrfToken());
-                
-                xhr.send(`productId=${product_id}&amountChange=${change}`);
-                // xhr.onload = function() {
-                //     if (this.status === 200) {
-                //         var response = JSON.parse(this.responseText);
-                //         document.getElementById('currentQuantity').textContent = response.newQuantity;
-                //         // document.getElementById('currentPrice').textContent = response.newQuantity * {{product.price}} +'원';
-                //         console.log('Update successful:', response.message);
-                //     } else {
-                //         console.error('Error updating:', this.statusText);
-                //     }
-                // };
-        }
-        function aa(element){
-            if (element.textContent ===  '+'){
-                a =  document.getElementById('currentQuantity').textContent
-                document.getElementById('currentQuantity').textContent = Number(a) + 1
-                document.getElementById('currentPrice').textContent = (Number(a)+1) * {{product.price}} +'원'
-            }
-            if (element.textContent ===  '-' && document.getElementById('currentQuantity').textContent>1 ){
-                    a =  document.getElementById('currentQuantity').textContent
-                    document.getElementById('currentQuantity').textContent = Number(a) - 1
-                    document.getElementById('currentPrice').textContent = (Number(a)-1) * {{product.price}} +'원'
-
-            }
-            } 
-            
-    </script>
-{% endblock %}
