--- product/templates/product/product_list.html
@@ -1,37 +1,0 @@
-{% extends 'product/base.html' %}
-
-
-    {% block main_area %}
-    <div class="container px-0 my-4" >
-        <div class="d-flex mx-auto">
-            <div class="m-2 font_size" style=" font-size: 20px;font-family: Lexend;font-weight: 700;line-height: 30px;">
-                <span style="font-size: 20px;font-family: Lexend;font-weight: 700;line-height: 30px;"></span>
-            </div>
-            
-            <div class="ms-3" >
-                <span class="row m-0 " style="font-size: 31px;font-family: Lexend;font-weight: 700;line-height: 48px;">{{ category }}</span>
-
-            </div>
-        </div>    
-    </div>
-
-
-    <div class="container">
-        
-        <div class="row col-sm-12 ms-auto">
-            {% for p in product_list %}
-            <div class="col-sm-3">
-                <div class=" mb-3">
-                    <div class="embed-responsive embed-responsive-4by3">
-                    <a href="{{ p.get_absolute_url }}"><img class="card-img-top embed-responsive-item" src="https://picsum.photos/seed/{{ p.id }}/800/200" width="150" height="180" alt="random_image" /></a>  
-                    </div> 
-                        <span class="badge text-bg-light ">브랜드명</span>
-                        <a href="{{ p.get_absolute_url }}" class="link-dark text-decoration-none"><h4 class="card-title h5">{{p.product_name}}</h4></a>
-                        <div class="d-flex justify-content-end">{{p.price}}원</div>          
-                </div>
-            </div>
-            {% endfor %}
-        </div>
-    </div>
-    {% endblock %}
-
