+++ seller_product/templates/product.html
@@ -58,6 +58,7 @@
                     </div> 
                         <span class="badge text-bg-light ">브랜드명</span>
                         <a href="/product/{{ p.product_id }}" class="link-dark text-decoration-none"><h4 class="card-title h5">{{p.product_name}}</h4></a>
+                        <p>{{p.product_description}}</p>
                         <div class="d-flex justify-content-end">{{p.price}}원</div>          
                 </div>
             </div>
