+++ seller_product/urls.py
@@ -5,13 +5,11 @@ from . import views
 app_name = 'product'
 
 urlpatterns = [
-    path('', views.ProductList.as_view()),
+    path('', views.ProductList.as_view(), name='product'),
     path('product/<int:pk>/', views.ProductDetail.as_view(), name='product_detail'),
     path('product/create/', views.ProductCreate.as_view(), name='product_create'),
     path('modify-cart/',views.modify_cart,name='modify_cart'),
     # path('<int:pk>/',views.ProductDetail.as_view()),
-    
-    
-    
-    # path('category/<str:slug>/',views.category_page)
+
+    path('category/<str:slug>/',views.category_page)
 ]
