+++ templates/base.html
@@ -19,8 +19,7 @@
 
   </head>
   <body>
-    <header>
-      <h1>Welcome to My Website</h1>
+    <header>      
       <nav class="navbar navbar-expand-lg navbar-white bg-white mb-3">
         <div class="container">
             <a class="navbar-brand" href="{% url 'home' %}"><span class="name_black">S</span><span class="name_color">E</span><span class="name_black">LLER</span></a>
