+++ templates/home.html
@@ -7,7 +7,7 @@
  {% block content %}
 <h1>HOME</h1>
 <p>Welcome to the homepage! Feel free to browse around.</p>
-<button><a href="{% url 'product' %}">상품</a></button>
+<button><a href="{% url 'product:product' %}">상품</a></button>
 {% endblock %} 
 {% block extra_js %}
